Concerts are already sorted
- Only Vault is converting items to Digests or Concerts. So make these conversions fileprivate. This way the code may safely assume that the [Concert] they use is already sorted. This way they do not need to sort them again!
- Reduces the transient number of Array<Concert> by about 1300 count by about 1 MB.
