--- Sources/MusicData/Artist+Digest.swift
@@ -1,35 +1,0 @@
-//
-//  Artist+Digest.swift
-//
-//
-//  Created by Greg Bolsinga on 9/6/23.
-//
-
-import Foundation
-
-extension Artist {
-  fileprivate func digest(
-    concerts: [Concert], lookup: Lookup, comparator: (Concert, Concert) -> Bool
-  ) -> ArtistDigest {
-    ArtistDigest(
-      artist: self,
-      shows: concerts.filter { $0.show.artists.contains(id) }.digests(comparator: comparator),
-      related: lookup.related(self),
-      firstSet: lookup.firstSet(artist: self),
-      spanRank: lookup.spanRank(artist: self),
-      showRank: lookup.showRank(artist: self),
-      venueRank: lookup.artistVenueRank(artist: self))
-  }
-}
-
-extension Array where Element == Artist {
-  public func digests(
-    concerts: [Concert], lookup: Lookup, comparator: (Concert, Concert) -> Bool
-  )
-    -> [ArtistDigest]
-  {
-    self.map {
-      $0.digest(concerts: concerts, lookup: lookup, comparator: comparator)
-    }
-  }
-}
